Stop truncating every character above U+FFFF
juce_wchar is 32 bits. The generator mapped it to uint32 by name, but libclang
resolves the alias to wchar_t before that mapping is consulted, and wchar_t was
mapped to uint16. Everything that returns a character went through the wrong
one.

It did not fail. String's operator[] returned 62976 for U+1F600, which is
0x1F600 cut to sixteen bits, and getAndAdvance did the same. A caller reading
anything outside the basic plane got a different character back and no
indication of it.

wchar_t is 32 bits on macOS and Linux, which are the platforms this binds.
JUCE only spells a type as wchar_t where wchar_t is 16-bit - Windows - and
CharPointer_UTF16 uses int16 there instead, so nothing on these platforms wants
the old mapping.

112 bindings change type. A test reads U+1F600 back through operator[] and
through the character pointer.

Found while writing coverage for CharacterFunctions, which is also where the
next defect showed up: isDigit and isWhitespace are each declared twice, once
for char and once for juce_wchar, and an argument from Nim converts to both, so
C++ cannot pick either. Those two are left out of the test with that written
down; the fix is a separate change.

+++ sources/june/juce_graphics.nim
@@ -921,7 +921,7 @@ proc getColourGlyphFormats*(this: Typeface): cint {.header: juce_graphics, impor
 proc setTypefaceCacheSize*(this: typedesc[Typeface], numFontsToCache: cint) {.header: juce_graphics, importcpp: "juce::Typeface::setTypefaceCacheSize(@)".}
 proc clearTypefaceCache*(this: typedesc[Typeface]) {.header: juce_graphics, importcpp: "juce::Typeface::clearTypefaceCache()".}
 proc scanFolderForFonts*(this: typedesc[Typeface], folder: File) {.header: juce_graphics, importcpp: "juce::Typeface::scanFolderForFonts(@)".}
-proc getNominalGlyphForCodepoint*(this: Typeface, arg1: uint16): CppOptional[uint32] {.header: juce_graphics, importcpp: "#.getNominalGlyphForCodepoint(@)".}
+proc getNominalGlyphForCodepoint*(this: Typeface, arg1: uint32): CppOptional[uint32] {.header: juce_graphics, importcpp: "#.getNominalGlyphForCodepoint(@)".}
 proc createSystemFallback*(this: Typeface, text: String, language: String): ReferenceCountedObjectPtr[Typeface] {.header: juce_graphics, importcpp: "#.createSystemFallback(@)".}
 proc findSystemTypeface*(this: typedesc[Typeface]): ReferenceCountedObjectPtr[Typeface] {.header: juce_graphics, importcpp: "juce::Typeface::findSystemTypeface()".}
 proc getSupportedFeatures*(this: Typeface): CppVector[FontFeatureTag] {.header: juce_graphics, importcpp: "#.getSupportedFeatures()".}
@@ -1095,8 +1095,8 @@ proc `AttributedStringAttribute=`*(this: var AttributedStringAttribute, arg1: va
 proc `==`*(this: AttributedStringAttribute, other: AttributedStringAttribute): bool {.error: "juce::AttributedString::Attribute defines no operator==; compare a property instead".}
 
 proc makePositionedGlyph*(): PositionedGlyph {.header: juce_graphics, importcpp: "juce::PositionedGlyph(@)".}
-proc makePositionedGlyph*(font: Font, character: uint16, glyphNumber: cint, anchorX: cfloat, baselineY: cfloat, width: cfloat, isWhitespace: bool): PositionedGlyph {.header: juce_graphics, importcpp: "juce::PositionedGlyph(@)".}
-proc getCharacter*(this: PositionedGlyph): uint16 {.header: juce_graphics, importcpp: "#.getCharacter()".}
+proc makePositionedGlyph*(font: Font, character: uint32, glyphNumber: cint, anchorX: cfloat, baselineY: cfloat, width: cfloat, isWhitespace: bool): PositionedGlyph {.header: juce_graphics, importcpp: "juce::PositionedGlyph(@)".}
+proc getCharacter*(this: PositionedGlyph): uint32 {.header: juce_graphics, importcpp: "#.getCharacter()".}
 proc isWhitespace*(this: PositionedGlyph): bool {.header: juce_graphics, importcpp: "#.isWhitespace()".}
 proc getLeft*(this: PositionedGlyph): cfloat {.header: juce_graphics, importcpp: "#.getLeft()".}
 proc getRight*(this: PositionedGlyph): cfloat {.header: juce_graphics, importcpp: "#.getRight()".}
